+++ owncloudComLibrary/src/main/java/com/owncloud/android/lib/resources/users/responses/UserInfoResponse.kt
@@ -21,7 +21,7 @@
  *   CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  *   THE SOFTWARE.
  */
-package com.owncloud.android.lib.resources.response
+package com.owncloud.android.lib.resources.users.responses
 
 import com.owncloud.android.lib.resources.users.RemoteUserInfo
 import com.squareup.moshi.Json
